requirements.txt
Flask==1.1.1
requests==2.23.0
python-dotenv==0.13.0
flask_cors==3.0.8
py3dbp==1.1.2